+++ src/components/feed/CreateFeed/FinishedStep.tsx
@@ -9,7 +9,7 @@ import {
 import { CreateFeedContext } from "./context";
 import { Types } from "./types";
 import { FaCogs } from "react-icons/fa";
-import { useAsync } from "../../../utils";
+import { useAsync } from "../../../api/common";
 import ReactJson from "react-json-view";
 
 const FinishedStep = () => {
